deployment/requirements-api.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
jinja2>=3.1.0
aiofiles>=23.0.0
pydantic>=2.0.0
requests>=2.31.0
